begin of refactor of fingerprint
